Rename lib package and move to same-name directory
